backend/requirements.txt
flask
flask_cors
pymongo
python-dotenv
requests
datetime